RIN-621 - Match popular Node file structures by moving custom functions into lib.
