+++ app.js
@@ -12,9 +12,9 @@ var http = require('http');
 var fs = require('fs');
 var nconf = require('nconf');
 var html = fs.readFileSync('./index.html');
-var token = require('./get-auth-token');
-var subscription = require('./subscription');
-var websocket = require('./websocket');
+var token = require('./lib/get-auth-token');
+var subscription = require('./lib/subscription');
+var websocket = require('./lib/websocket');
 
 /**
  * Setup nconf to use (in-order):
