+++ lib/get-auth-token.js
