+++ lib/websocket.js
@@ -54,8 +54,10 @@ function newWebSocket(response) {
       // @todo: Ping third party with message data.
       // http://rinnai.mcdev/api/dss/VXRINNAI0000001/WH1_INF_UNT_ECC?_format=json
       // http://rinnai.mcdev/api/dss/{dsn}/{property_name}?_format=json
+      // need to pass updated date as well.
       console.log(`dsn: ${ messageInfo.metadata.dsn }`);
       console.log(`property_name: ${ messageInfo.metadata.property_name }`);
+      console.log(`data_updated_at: ${ messageInfo.datapoint.updated_at }`);
     }
   });
